% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_seplyr_pipe.R
\name{insert_seplyr_pipe}
\alias{insert_seplyr_pipe}
\title{Inserts a Seplyr pipe.}
\usage{
insert_seplyr_pipe()
}
\value{
\code{"\%.>\%"}
}
\description{
#' Assign this to a shortcut in Options > Code > Modify Keyboard Shortcuts
}
\examples{
insert_seplyr_pipe <- function() {
    rstudioapi::insertText(
        location = rstudioapi::document_position(0, 0),
        text = ' \%.>\% '
    )
}
}
