% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint_doc.R
\name{lint_doc}
\alias{lint_doc}
\title{Lint an open file in the editor}
\usage{
lint_doc()
}
\value{
None
}
\description{
Assign this to a shortcut in Options > Code > Modify Keyboard Shortcuts
}
\examples{
lint_doc <- function() {
    lintr::lint(
        rstudioapi::getSourceEditorContext()$path
    )
}

}
